+++ NEWS.md
@@ -6,7 +6,12 @@
 #### Highlights
 * Update ScalePar() function to properly handle recession coefficients
 * Add ScaleAquiferData() function to scale RETRATE recession coefficient
-* Add example AquiferData.txt file to demo model; taken from HYPE wiki: http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt
+* Add example AquiferData.txt and FloodData.txt files to demo model; taken from HYPE wiki:
+  * http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:aquiferdata.txt
+  * http://www.smhi.net/hype/wiki/doku.php?id=start:hype_file_reference:flooddata.txt
+* Add ScaleFloodData() function to scale recession cofficients
+* Add ReadFloodData() and WriteFloodData() functions
+* Update CompareFiles() function to warn if "type" argument is not in the supported list of options
 
 ## HYPEtools 1.6.1 (2924-01-12)
 *Enhancements and bug fixes*